+++ ios/platform/InMobiSDK.xcframework/ios-arm64_armv7/InMobiSDK.framework/Headers/InMobiSDK.h
@@ -1,0 +1,46 @@
+//
+//  InMobiSDK.h
+//  InMobiSDK
+//
+//  Copyright © 2016 InMobi. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+#ifndef InMobiSDK_h
+#define InMobiSDK_h
+
+//! Project version number for InMobiSDK.
+FOUNDATION_EXPORT double InMobiSDKVersionNumber;
+
+//! Project version string for InMobiSDK.
+FOUNDATION_EXPORT const unsigned char InMobiSDKVersionString[];
+
+#import <InMobiSDK/IMSdk.h>
+#import <InMobiSDK/IMCommonConstants.h>
+#import <InMobiSDK/IMRequestStatus.h>
+
+#import <InMobiSDK/IMBanner.h>
+#import <InMobiSDK/IMBannerDelegate.h>
+#import <InMobiSDK/IMBannerPreloadManager.h>
+
+#import <InMobiSDK/IMInterstitial.h>
+#import <InMobiSDK/IMInterstitialDelegate.h>
+#import <InMobiSDK/IMInterstitialPreloadManager.h>
+
+#import <InMobiSDK/IMNative.h>
+#import <InMobiSDK/IMNativeDelegate.h>
+
+#import <InMobiSDK/IMCoreConstants.h>
+#import <InMobiSDK/IMAdMetaInfo.h>
+
+#if __has_include(<InMobiMediationSDK/InMobiMediationSDK.h>)
+#import <InMobiMediationSDK/InMobiMediationSDK.h>
+#endif
+
+#import <InMobiSDK/IMUnifiedIdService.h>
+#import <InMobiSDK/IMUnifiedIdDelegate.h>
+#import <InMobiSDK/IMUserDataModel.h>
+#import <InMobiSDK/IMUserDataTypes.h>
+
+#endif /* InMobiSDK_h */